+++ app/src/main/java/com/example/zoomrad/presentation/screens/tabs/monitor/MonitoringScreen.kt
@@ -1,15 +1,34 @@
 package com.example.zoomrad.presentation.screens.tabs.monitor
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.GridItemSpan
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
@@ -21,7 +40,8 @@ import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.presenter.vm.transaction.TransactionViewModel
 import java.text.SimpleDateFormat
-import java.util.*
+import java.util.Date
+import java.util.Locale
 
 data class MonitoringTransactionData(
     val id: String,
@@ -59,11 +79,11 @@ fun MonitoringScreen(
             id = api.id,
             title = api.type.replace("_", " "),
             subtitle = api.description,
-            amount = amountPrefix + String.format(Locale.getDefault(), "%,d", api.amount).replace(',', ' '),
+            amount = amountPrefix + String.format(Locale.getDefault(), "%,d", api.amount.toLong()).replace(',', ' '),
             time = timeOutputFormat.format(dateObj ?: Date()),
             isIncome = isIncome,
             date = dateOutputFormat.format(dateObj ?: Date()),
-            rawAmount = api.amount
+            rawAmount = api.amount.toLong()
         )
     }
 
@@ -100,14 +120,14 @@ fun MonitoringScreen(
             } else {
                 item {
                     SummaryCard(
-                        amount = String.format(Locale.getDefault(), "%,d", totalIncome).replace(',', ' '),
+                        amount = String.format(Locale.getDefault(), "%,d", totalIncome.toLong()).replace(',', ' '),
                         label = "Tushumlar",
                         isIncome = true
                     )
                 }
                 item {
                     SummaryCard(
-                        amount = String.format(Locale.getDefault(), "%,d", totalExpense).replace(',', ' '),
+                        amount = String.format(Locale.getDefault(), "%,d", totalExpense.toLong()).replace(',', ' '),
                         label = "Chiqimlar",
                         isIncome = false
                     )
